refactor: restructure LLM modules and improve model handling
- Move LLM-related files to dedicated directories (llm/ and models/)
- Refactor compactTask and generateTaskTitle to use ai-sdk directly
- Create new stepAgent function for streaming text with tools
- Refactor model creation into separate model files
- Remove legacy LLM request handling code
- Improve PersistManager for graceful shutdown

This reorganization improves the modularity of LLM handling and prepares for future enhancements.
