+++ packages/livekit/src/chat/models/pochi.ts
@@ -1,0 +1,56 @@
+import type { LanguageModelV2 } from "@ai-sdk/provider";
+import { EventSourceParserStream } from "@ai-sdk/provider-utils";
+import type { RequestData } from "../../types";
+
+export function createPochiModel(
+  id: string | undefined,
+  llm: Extract<RequestData["llm"], { type: "pochi" }>,
+) {
+  return {
+    specificationVersion: "v2",
+    provider: "pochi",
+    modelId: llm.modelId || "<default>",
+    // FIXME(meng): fill supported urls based on modelId.
+    supportedUrls: {},
+    doGenerate: async () => Promise.reject("Not implemented"),
+    doStream: async ({ prompt, abortSignal, stopSequences, tools }) => {
+      const resp = await llm.apiClient.api.chat.stream.$post(
+        {
+          json: {
+            id,
+            model: llm.modelId,
+            callOptions: {
+              prompt,
+              stopSequences,
+              tools,
+            },
+          },
+        },
+        {
+          init: {
+            signal: abortSignal,
+          },
+        },
+      );
+
+      if (!resp.ok || !resp.body) {
+        throw new Error(`Failed to fetch: ${resp.status} ${resp.statusText}`);
+      }
+
+      const stream = resp.body
+        .pipeThrough(new TextDecoderStream())
+        .pipeThrough(new EventSourceParserStream())
+        .pipeThrough(
+          new TransformStream({
+            async transform({ data }, controller) {
+              if (data === "[DONE]") {
+                return;
+              }
+              controller.enqueue(JSON.parse(data));
+            },
+          }),
+        );
+      return { stream };
+    },
+  } satisfies LanguageModelV2;
+}
